Handle file parameters on pipeline builds
Fixes [JENKINS-27413] and [JENKINS-47333] (duplicate). Adds file
parameter support on pipeline jobs, while maintaining compatibility
with freestyle projects.

WorkflowJob does not trigger createBuildWrapper during the job run, so
the side effect (copying the file) was not performed on a pipeline job.

My proposal consists on applying the side effect in the buildEnvironment
method, which is invoked both by a Pipeline and a Freestyle job.

I've added a variable to keep track of the internal state, so that the
copying is only performed once.

Also, modified doDynamic() method to allow display of the fileParameter
when inspecting the run (after completion) by casting to the parent
class.

I've also fixed the related broken tests. I had to add a custom
workspace so that buildEnvironment() would be triggered, and the side
effect executed, other than that, the logic remains unchanged.